modify ADD and AND instructions in order to pass the tests
in the last commit the ADD function received a copy of the registers array, which wasn't correct because in that case the changes made to the array inside the function weren't applied in the original array. To solve this, the ADD function now receives a mutable reference to the registers array. So does the AND function